refactor: 쿠폰 발급 스케줄러의 패키지 위치 수정 및 표현 계층 패키징 세분화
